Add reflection

+++ Assets/Shaders/AddShader.shader
@@ -40,7 +40,7 @@ Shader "Hidden/AddShader"
             sampler2D _MainTex;
             float uSample;
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 return float4(tex2D(_MainTex, i.uv).rgb, 1.0f / (uSample + 1.0f));
             }
